+++ topydo/commands/TagCommand.py
@@ -15,11 +15,14 @@
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
 from topydo.lib.Command import Command, InvalidCommandArgument
-from topydo.lib.PrettyPrinterFilter import PrettyPrinterNumbers
+from topydo.lib.Config import config
+from topydo.lib.prettyprinters.Numbers import PrettyPrinterNumbers
+from topydo.lib.RelativeDate import relative_date_to_date
 from topydo.lib.TodoListBase import InvalidTodoException
 
+
 class TagCommand(Command):
-    def __init__(self, p_args, p_todolist,
+    def __init__(self, p_args, p_todolist, #pragma: no branch
                  p_out=lambda a: None,
                  p_err=lambda a: None,
                  p_prompt=lambda a: None):
@@ -67,7 +70,7 @@ class TagCommand(Command):
 
     def _choose(self):
         """
-        Returns the chosen number of the tag value to process (or "all")
+        Returns the chosen number of the tag value to process (or "all").
         """
         answer = "all"
 
@@ -89,7 +92,16 @@ class TagCommand(Command):
 
         return answer
 
+    def _convert_relative_dates(self):
+        if self.tag == config().tag_start() or self.tag == config().tag_due():
+            real_date = relative_date_to_date(self.value)
+
+            if real_date:
+                self.value = real_date.isoformat()
+
     def _set_helper(self, p_old_value=""):
+        self._convert_relative_dates()
+
         old_src = self.todo.source()
         self.todo.set_tag(self.tag, self.value, self.force_add, p_old_value)
 
@@ -103,7 +115,7 @@ class TagCommand(Command):
             if answer == "all":
                 for value in self.current_values:
                     self._set_helper(value)
-            elif answer != None and self.value != self.current_values[answer]:
+            elif answer is not None and self.value != self.current_values[answer]:
                 self._set_helper(self.current_values[answer])
 
         else:
@@ -130,6 +142,6 @@ is omitted, the tag is removed from the todo item.
 
 -a : Do not change the current value of the tag if it exists, but add a new
      value.
--f : Force setting/removing all values of the tag. Prevents interaction with the
-     user.
+-f : Force setting/removing all values of the tag. Prevents interaction with
+     the user.
 """