+++ topydo/lib/Colors.py
@@ -18,7 +18,8 @@
 
 from topydo.lib.Config import config
 
-NEUTRAL_COLOR  = '\033[0m'
+NEUTRAL_COLOR = '\033[0m'
+
 
 class Colors(object):
     def __init__(self):
@@ -48,13 +49,13 @@ class Colors(object):
 
         try:
             if p_safe:
-                if 8 > int(p_int) >=0:
+                if 8 > int(p_int) >= 0:
                     return '\033[{};3{}m'.format(decoration, str(p_int))
                 elif 16 > int(p_int):
                     p_int = int(p_int) - 8
                     return '\033[{};1;3{}m'.format(decoration, str(p_int))
 
-            if 256 > int(p_int) >=0:
+            if 256 > int(p_int) >= 0:
                 return '\033[{};38;5;{}m'.format(decoration, str(p_int))
             else:
                 return NEUTRAL_COLOR