+++ topydo/lib/Config.py
@@ -14,9 +14,9 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
+import configparser
 import os
-
-from six.moves import configparser
+import shlex
 
 class ConfigError(Exception):
     def __init__(self, p_text):
@@ -24,6 +24,7 @@ class ConfigError(Exception):
 
     def __str__(self):
         return self.text
+
 
 class _Config:
     def __init__(self, p_path=None, p_overrides=None):
@@ -40,6 +41,7 @@ class _Config:
         """
         self.sections = [
             'add',
+            'aliases',
             'colorscheme',
             'dep',
             'ls',
@@ -49,46 +51,72 @@ class _Config:
         ]
 
         self.defaults = {
-            # topydo
-            'default_command': 'ls',
-            'colors': '1',
-            'filename' : 'todo.txt',
-            'archive_filename' : 'done.txt',
-            'identifiers': 'linenumber',
+            'topydo': {
+                'default_command': 'ls',
+                'colors': '1',
+                'filename': 'todo.txt',
+                'archive_filename': 'done.txt',
+                'identifiers': 'linenumber',
+                'backup_count': '5',
+            },
 
-            # add
-            'auto_creation_date': '1',
+            'add': {
+                'auto_creation_date': '1',
+            },
 
-            # ls
-            'hide_tags': 'id,p,ical',
-            'indent': 0,
-            'list_limit': '-1',
+            'ls': {
+                'hide_tags': 'id,p,ical',
+                'indent': '0',
+                'list_limit': '-1',
+                'list_format': '|%I| %x %{(}p{)} %c %s %k %{due:}d %{t:}t',
+            },
 
-            # tags
-            'tag_start': 't',
-            'tag_due': 'due',
-            'tag_star': 'star',
+            'tags': {
+                'tag_start': 't',
+                'tag_due': 'due',
+                'tag_star': 'star',
+            },
 
-            # sort
-            'keep_sorted': '0',
-            'sort_string': 'desc:importance,due,desc:priority',
-            'ignore_weekends': '1',
+            'sort': {
+                'keep_sorted': '0',
+                'sort_string': 'desc:importance,due,desc:priority',
+                'ignore_weekends': '1',
+            },
 
-            # dep
-            'append_parent_projects': '0',
-            'append_parent_contexts': '0',
+            'dep': {
+                'append_parent_projects': '0',
+                'append_parent_contexts': '0',
+            },
 
-            # colorscheme
-            'project_color': 'red',
-            'context_color': 'magenta',
-            'metadata_color': 'green',
-            'link_color': 'cyan',
-            'priority_colors': 'A:cyan,B:yellow,C:blue',
+            'colorscheme': {
+                'project_color': 'red',
+                'context_color': 'magenta',
+                'metadata_color': 'green',
+                'link_color': 'cyan',
+                'priority_colors': 'A:cyan,B:yellow,C:blue',
+            },
+
+            'aliases': {
+                'lsproj': 'lsprj',
+                'listprj': 'lsprj',
+                'listproj': 'lsprj',
+                'listproject': 'lsprj',
+                'listprojects': 'lsprj',
+                'listcon': 'lscon',
+                'listcontext': 'lscon',
+                'listcontexts': 'lscon',
+            },
         }
 
         self.config = {}
 
-        self.cp = configparser.ConfigParser(self.defaults)
+        self.cp = configparser.RawConfigParser()
+
+        for section in self.defaults:
+            self.cp.add_section(section)
+
+            for option, value in self.defaults[section].items():
+                self.cp.set(section, option, value)
 
         files = [
             "/etc/topydo.conf",
@@ -100,11 +128,10 @@ class _Config:
 
         # when a path is given, *only* use the values in that file, or the
         # defaults listed above.
-        if p_path != None:
+        if p_path is not None:
             files = [p_path]
 
         self.cp.read(files)
-
         self._supplement_sections()
 
         if p_overrides:
@@ -126,7 +153,7 @@ class _Config:
         try:
             return self.cp.getboolean('topydo', 'colors')
         except ValueError:
-            return self.defaults['colors'] == '1'
+            return self.defaults['topydo']['colors'] == '1'
 
     def todotxt(self):
         return os.path.expanduser(self.cp.get('topydo', 'filename'))
@@ -137,23 +164,32 @@ class _Config:
     def identifiers(self):
         return self.cp.get('topydo', 'identifiers')
 
+    def backup_count(self):
+        try:
+            value = self.cp.getint('topydo', 'backup_count')
+            if value < 0:
+                value = 0
+            return value
+        except ValueError:
+            return int(self.defaults['topydo']['backup_count'])
+
     def list_limit(self):
         try:
             return self.cp.getint('ls', 'list_limit')
         except ValueError:
-            return int(self.defaults['list_limit'])
+            return int(self.defaults['ls']['list_limit'])
 
     def list_indent(self):
         try:
             return self.cp.getint('ls', 'indent')
         except ValueError:
-            return int(self.defaults['indent'])
+            return int(self.defaults['ls']['indent'])
 
     def keep_sorted(self):
         try:
             return self.cp.getboolean('sort', 'keep_sorted')
         except ValueError:
-            return self.defaults['keep_sorted'] == '1'
+            return self.defaults['sort']['keep_sorted'] == '1'
 
     def sort_string(self):
         return self.cp.get('sort', 'sort_string')
@@ -162,19 +198,19 @@ class _Config:
         try:
             return self.cp.getboolean('sort', 'ignore_weekends')
         except ValueError:
-            return self.defaults['ignore_weekends'] == '1'
+            return self.defaults['sort']['ignore_weekends'] == '1'
 
     def append_parent_projects(self):
         try:
             return self.cp.getboolean('dep', 'append_parent_projects')
         except ValueError:
-            return self.defaults['append_parent_projects'] == '1'
+            return self.defaults['dep']['append_parent_projects'] == '1'
 
     def append_parent_contexts(self):
         try:
             return self.cp.getboolean('dep', 'append_parent_contexts')
         except ValueError:
-            return self.defaults['append_parent_contexts'] == '1'
+            return self.defaults['dep']['append_parent_contexts'] == '1'
 
     def _get_tag(self, p_tag):
         try:
@@ -197,10 +233,13 @@ class _Config:
         """ Returns a list of tags to be hidden from the 'ls' output. """
         hidden_tags = self.cp.get('ls', 'hide_tags')
         # pylint: disable=no-member
-        return [] if hidden_tags == '' else hidden_tags.split(',')
+        return [] if hidden_tags == '' else [tag.strip() for tag in
+                                             hidden_tags.split(',')]
 
     def priority_colors(self):
-        """ Returns a dict with priorities as keys and color numbers as value. """
+        """
+        Returns a dict with priorities as keys and color numbers as value.
+        """
         pri_colors_str = self.cp.get('colorscheme', 'priority_colors')
 
         def _str_to_dict(p_string):
@@ -213,11 +252,11 @@ class _Config:
 
         try:
             if pri_colors_str == '':
-                pri_colors_dict = {'A':'', 'B': '', 'C': ''}
+                pri_colors_dict = {'A': '', 'B': '', 'C': ''}
             else:
                 pri_colors_dict = _str_to_dict(pri_colors_str)
         except ValueError:
-            pri_colors_dict = _str_to_dict(self.defaults['priority_colors'])
+            pri_colors_dict = _str_to_dict(self.defaults['colorscheme']['priority_colors'])
 
         return pri_colors_dict
 
@@ -225,31 +264,52 @@ class _Config:
         try:
             return self.cp.get('colorscheme', 'project_color')
         except ValueError:
-            return int(self.defaults['project_color'])
+            return int(self.defaults['colorscheme']['project_color'])
 
     def context_color(self):
         try:
             return self.cp.get('colorscheme', 'context_color')
         except ValueError:
-            return int(self.defaults['context_color'])
+            return int(self.defaults['colorscheme']['context_color'])
 
     def metadata_color(self):
         try:
             return self.cp.get('colorscheme', 'metadata_color')
         except ValueError:
-            return int(self.defaults['metadata_color'])
+            return int(self.defaults['colorscheme']['metadata_color'])
 
     def link_color(self):
         try:
             return self.cp.get('colorscheme', 'link_color')
         except ValueError:
-            return int(self.defaults['link_color'])
+            return int(self.defaults['colorscheme']['link_color'])
 
     def auto_creation_date(self):
         try:
             return self.cp.getboolean('add', 'auto_creation_date')
         except ValueError:
-            return self.defaults['auto_creation_date'] == '1'
+            return self.defaults['add']['auto_creation_date'] == '1'
+
+    def aliases(self):
+        """
+        Returns dict with aliases names as keys and pairs of actual
+        subcommand and alias args as values.
+        """
+        aliases = self.cp.items('aliases')
+        alias_dict = dict()
+
+        for alias, meaning in aliases:
+            meaning = shlex.split(meaning)
+            real_subcommand = meaning[0]
+            alias_args = meaning[1:]
+            alias_dict[alias] = (real_subcommand, alias_args)
+
+        return alias_dict
+
+    def list_format(self):
+        """ Returns the list format used by `ls` """
+        return self.cp.get('ls', 'list_format')
+
 
 def config(p_path=None, p_overrides=None):
     """
@@ -263,7 +323,7 @@ def config(p_path=None, p_overrides=None):
     passed value instead. Structure: (section, option) => value
     The previous configuration instance will be discarded.
     """
-    if not config.instance or p_path != None or p_overrides != None:
+    if not config.instance or p_path is not None or p_overrides is not None:
         try:
             config.instance = _Config(p_path, p_overrides)
         except configparser.ParsingError as perr: