+++ topydo/lib/MultiCommand.py
@@ -14,17 +14,16 @@
 # You should have received a copy of the GNU General Public License
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 
-from six import u
-
 from topydo.lib.ExpressionCommand import ExpressionCommand
 from topydo.lib.TodoListBase import InvalidTodoException
+
 
 class MultiCommand(ExpressionCommand):
     """
     A common class for operations that can work with multiple todo IDs.
     """
 
-    def __init__(self, p_args, p_todolist,
+    def __init__(self, p_args, p_todolist, #pragma: no branch
                  p_out=lambda a: None,
                  p_err=lambda a: None,
                  p_prompt=lambda a: None):
@@ -42,7 +41,7 @@ class MultiCommand(ExpressionCommand):
 
     def process_flag(self, p_option, p_value):
         """ Default implementation of processing specific flags. """
-        pass
+        raise NotImplementedError
 
     def _process_flags(self):
         opts, long_opts = self.get_flags()
@@ -62,7 +61,7 @@ class MultiCommand(ExpressionCommand):
         self.todos = self._view().todos
 
     def get_todos(self):
-        """ Gets todo objects from supplied todo IDs """
+        """ Gets todo objects from supplied todo IDs. """
         if self.is_expression:
             self.get_todos_from_expr()
         else:
@@ -79,17 +78,17 @@ class MultiCommand(ExpressionCommand):
 
     def _catch_todo_errors(self):
         """
-        Returns None or list of error messages depending on number of valid todo
-        objects and number of invalid todo IDs.
+        Returns None or list of error messages depending on number of valid
+        todo objects and number of invalid todo IDs.
 
-        In case of multiple invalid todo IDs we generate separate error message for each
-        one of them with information about supplied ID.
+        In case of multiple invalid todo IDs we generate separate error message
+        for each one of them with information about supplied ID.
         """
         errors = []
 
         if len(self.invalid_numbers) > 1 or len(self.invalid_numbers) > 0 and len(self.todos) > 0:
             for number in self.invalid_numbers:
-                errors.append(u("Invalid todo number given: {}.").format(number))
+                errors.append(u"Invalid todo number given: {}.".format(number))
         elif len(self.invalid_numbers) == 1 and len(self.todos) == 0:
             errors.append("Invalid todo number given.")
         elif len(self.todos) == 0 and len(self.invalid_numbers) == 0:
@@ -105,14 +104,14 @@ class MultiCommand(ExpressionCommand):
         Operations specific for particular command dealing with multiple todo
         IDs.
         """
-        pass
+        raise NotImplementedError
 
     def _execute_not_multi(self):
         """
         Some commands can do something else besides operating on multiple todo
         IDs. This method is a wrapper for those other operations.
         """
-        pass
+        raise NotImplementedError
 
     def execute(self):
         if not super(MultiCommand, self).execute():