+++ topydo/lib/Utils.py
@@ -1,5 +1,5 @@
 # Topydo - A todo.txt client written in Python.
-# Copyright (C) 2014 Bram Schoenmakers <me@bramschoenmakers.nl>
+# Copyright (C) 2014 - 2015 Bram Schoenmakers <me@bramschoenmakers.nl>
 #
 # This program is free software: you can redistribute it and/or modify
 # it under the terms of the GNU General Public License as published by
@@ -18,8 +18,17 @@
 Various utility functions.
 """
 
-from datetime import date
 import re
+
+from collections import namedtuple
+from datetime import date
+
+# shutil.get_terminal_size was added to the standard library in Python 3.3
+try:
+    from shutil import get_terminal_size as _get_terminal_size  # pylint: disable=no-name-in-module
+except ImportError:
+    from backports.shutil_get_terminal_size import get_terminal_size as _get_terminal_size  # pylint: disable=import-error
+
 
 def date_string_to_date(p_date):
     """
@@ -32,19 +41,39 @@ def date_string_to_date(p_date):
         parsed_date = re.match(r'(\d{4})-(\d{2})-(\d{2})', p_date)
         if parsed_date:
             result = date(
-                int(parsed_date.group(1)), # year
-                int(parsed_date.group(2)), # month
-                int(parsed_date.group(3))  # day
+                int(parsed_date.group(1)),  # year
+                int(parsed_date.group(2)),  # month
+                int(parsed_date.group(3))   # day
             )
         else:
             raise ValueError
 
     return result
 
+
 def is_valid_priority(p_priority):
-    return p_priority != None and re.match(r'^[A-Z]$', p_priority) != None
+    return p_priority is not None and re.match(r'^[A-Z]$', p_priority) is not None
+
 
 def escape_ansi(p_string):
     return escape_ansi.pattern.sub('', p_string)
 
 escape_ansi.pattern = re.compile(r'\x1b[^m]*m')
+
+
+def get_terminal_size():
+    """
+    Try to determine terminal size at run time. If that is not possible,
+    returns the default size of 80x24.
+    """
+    try:
+        sz = _get_terminal_size()
+    except ValueError:
+        """
+        This can result from the 'underlying buffer being detached', which
+        occurs during running the unittest on Windows (but not on Linux?)
+        """
+        terminal_size = namedtuple('Terminal_Size', 'columns lines')
+        sz = terminal_size((80, 24))
+
+    return sz